Remove 'nightly' benchmarking of internal modules
Instead just bench the public .svd() method along
with the other benchmarks.
