+++ Cargo.toml
@@ -16,6 +16,3 @@ matrixmultiply = "0.1.8"
 
 [dev-dependencies]
 rand = "0.3"
-
-[features]
-nightly = []
