+++ benches/lib.rs
@@ -2,7 +2,9 @@
 
 extern crate rulinalg;
 extern crate test;
+extern crate rand;
 
 mod linalg {
 	mod matrix;
+	mod svd;
 }
